requirements.txt
# concrete-ml 僅支援 Python 3.9
concrete-ml==1.0.0
numpy>=1.21